requirements.txt
pydantic
pyautogui
qdrant-client
pymongo
openai
google-generativeai
python-dotenv
mss
pytesseract
